refactor: extract AssistantStatus and InteractionType from AppDelegate.swift
